Changed the rest of the tutorials and the amazon network
