+++ pycascades/amazon/__init__.py
@@ -1,0 +1,5 @@
+
+
+from . import amazon
+
+from pycascades.amazon.amazon import generate_network
